requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
beautifulsoup4==4.14.2
blis==1.3.0
catalogue==2.0.10
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpathlib==0.23.0
colorama==0.4.6
comtypes==1.4.13
confection==0.1.5
cymem==2.0.11
filelock==3.20.0
fsspec==2025.10.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
llvmlite==0.45.1
MarkupSafe==3.0.3
murmurhash==1.0.13
numba==0.62.1
numpy==2.3.4
packaging==25.0
pefile==2023.2.7
preshed==3.0.10
PyAudio==0.2.14
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pypiwin32==223
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.1
python-dotenv==1.2.1
pyttsx3==2.99
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
resampy==0.4.3
safetensors==0.6.2
smart_open==7.4.4
soundfile==0.13.1
soupsieve==2.8
spacy==3.8.8
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
thinc==8.3.8
tokenizers==0.22.1
tqdm==4.67.1
transformers==4.57.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.2
wikipedia==1.4.0
wrapt==2.0.1
